# back-end/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Good Bank
  version: 1.0.0
servers:
  - url: http://159.89.47.38:5051
tags:
  - name: Record
    description: Operations related to records
paths:
  /record:
    post:
      tags:
        - Record
      summary: Create a new record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Created
        '200':
          description: Success
    get:
      tags:
        - Record
      summary: Get all customer records
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                required:
                  - _id
                items: 
                  allOf:
                    - type: object
                      properties:
                        _id:
                          type: string
                    - $ref: '#/components/schemas/Customer'

  /record/{email}:
    get:
      tags:
        - Record
      summary: Get a record by email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      _id:
                        type: string
                  - $ref: '#/components/schemas/Customer'

  /record/{_id}:
    patch:
      tags:
        - Record
      summary:
        Update a record by ID
      parameters:
        - name: _id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: OK                  
    delete:
      tags:
        - Record
      summary:
        Delete a record by ID
      parameters:
        - name: _id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '200':
          description: Success
          
components:
  schemas:
    Customer:
      type: object
      required:
        - name
        - email
        - password
        - balance
      properties:
        name:
          type: string
          example: Anthony Demott
        email:
          type: string
          example: charchar@pokecenter.edu
        password:
          type: string
          example: dragonrage123
        balance:
          type: integer
          example: 50
        

